# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Backend Core API",
        "description": "Learning Management System Backend API Documentation. This API provides comprehensive endpoints for managing courses, users, enrollments, and learning progress.",
        "contact": {
            "name": "LMS API Support",
            "email": "api-support@lms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LMS API Server"
        }
    ],
    "paths": {
...
        "/courses/{id}/enroll": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Enroll in a course",
                "description": "Enroll the authenticated user in a course",
                "operationId": "7abbd560f122a0d11c551238ecab6b87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrolled in course successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enrolled in course successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already enrolled or course not available"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/courses/{courseId}/enroll": {
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll in a course",
                "description": "Enroll the authenticated user in a specific course",
                "operationId": "2a4bed1da07f472ffa16129ca837388c",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully enrolled in course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully enrolled in course"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "enrolled_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "progress_percentage": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0
                                                },
                                                "course": {
                                                    "$ref": "#/components/schemas/Course"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already enrolled or course not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You are already enrolled in this course"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get user enrollments",
                "description": "Get all enrollments for the authenticated user",
                "operationId": "6a2d59a345387de602a0cbcae9ccb173",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by enrollment status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "completed",
                                "suspended",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrollments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enrollments retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "active"
                                                            },
                                                            "enrolled_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "completed_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "nullable": true
                                                            },
                                                            "progress_percentage": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 45.5
                                                            },
                                                            "last_accessed_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "nullable": true
                                                            },
                                                            "course": {
                                                                "$ref": "#/components/schemas/Course"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "title": "Course",
                "description": "Course model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "slug": {
                        "type": "string",
                        "example": "introduction-to-laravel"
                    },
                    "short_description": {
                        "type": "string",
                        "example": "Learn Laravel framework basics"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comprehensive course covering Laravel fundamentals..."
                    },
                    "thumbnail_url": {
                        "type": "string",
                        "example": "https://example.com/thumbnails/course1.jpg"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "beginner",
                            "intermediate",
                            "advanced"
                        ],
                        "example": "beginner"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "estimated_duration_hours": {
                        "type": "integer",
                        "example": 20
                    },
                    "language": {
                        "type": "string",
                        "example": "en"
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_free": {
                        "type": "boolean",
                        "example": false
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "instructor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "enrollment_count": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_chapters": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}